Added monday-first weeks option at calendar
I've added an option to allow different a calendar configuration, since Monday is the first day of week in some countries. The problem is that tricky that I found at
http://apple.stackexchange.com/questions/93906/cal-command-start-monday

It's really awful but it seems to be the best choice since cal doesn't have any configuration setting to allow this.